examples/wdl/requirements.txt
miniwdl==1.4.3
python-pachyderm==7.1.1
